update the web site security

+++ be-events/delete-event.php
@@ -6,7 +6,7 @@ $event_id = $_POST['id'] ?? null;
 if($_SERVER['REQUEST_METHOD'] === 'POST'){
     $user_id = $_SESSION['user_id'];
 
-    $query = "DELETE FROM `events` WHERE `events`.`id` = $event_id";
+    $query = "DELETE FROM `events` WHERE `events`.`id` = $event_id && `user_id` = $user_id";
 
     $result = mysqli_query($conn, $query);
     if($result){
